requirements.txt
aiohttp==3.5.4
aiomqtt==0.1.1
aioredis==1.2.0
APScheduler==3.6.0
astral==1.10.1
async-timeout==3.0.1
attrs==19.1.0
chardet==3.0.4
Click==7.0
Cython==0.29.6
dnspython==1.16.0
docopt==0.6.2
eventlet==0.24.1
gmqtt==0.3.1
greenlet==0.4.15
hbmqtt==0.9.5
hiredis==1.0.0
idna==2.8
idna-ssl==1.1.0
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
monotonic==1.5
multidict==4.5.2
paho-mqtt==1.4.0
passlib==1.7.1
psutil==5.6.1
pyserial==3.4
python-engineio==3.5.0
python-socketio==3.1.2
pytz==2018.9
PyYAML==5.1
redis==3.2.1
six==1.12.0
transitions==0.6.9
typing-extensions==3.7.2
tzlocal==1.5.1
websockets==7.0
Werkzeug==0.15.2
XBee==2.3.2
yarl==1.3.0
